* 'develop' of gitlab.com:imazeapp/andorid_nevo:
  [change name add include layout]
  [fixing] [according BLE R22, add new fields to steps table]
  [change color and add string arrary resoures]
  [fixing] [fix user profile edit screen,save it to local database and sync to watch]
  [complete solar pie chart color]

# Conflicts:
#	Programming/Nevo/app/app.iml
#	Programming/Nevo/app/build.gradle
#	Programming/Nevo/app/src/androidTest/java/com/medcorp/nevo/database/AlarmDatabaseHelperTest.java
#	Programming/Nevo/app/src/main/java/com/medcorp/fragment/LunarMainSleepFragment.java
#	Programming/Nevo/app/src/main/java/com/medcorp/fragment/LunarMainSolarFragment.java
#	Programming/Nevo/app/src/main/java/com/medcorp/view/graphs/MainStepsBarChart.java
#	Programming/Nevo/app/src/main/res/layout/lunar_main_sleep_fragment_layout.xml

+++ Programming/Nevo/app/src/main/java/com/medcorp/model/User.java
@@ -8,7 +8,7 @@ public class User {
 
     private int age = 18;
 
-    private float weight = 77;//kg
+    private int weight = 77;//kg
 
     private int height = 175;//cm
 
@@ -56,7 +56,7 @@ public class User {
         this.age = age;
     }
 
-    public void setWeight(float weight) {
+    public void setWeight(int weight) {
         this.weight = weight;
     }
 
@@ -80,7 +80,7 @@ public class User {
         return age;
     }
 
-    public float getWeight() {
+    public int getWeight() {
         return weight;
     }
